put DTOs in separate packages
