+++ applicant-server/src/main/kotlin/ut/isep/interview/api/SectionController.kt
@@ -1,6 +1,6 @@
 package ut.isep.interview.api
 
-import dto.SectionReadDTO
+import dto.section.SectionReadDTO
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.Schema
