+++ management-server/src/main/kotlin/ut/isep/management/controller/ApplicantController.kt
@@ -1,6 +1,11 @@
 package ut.isep.management.controller
 
 import dto.*
+import dto.applicant.ApplicantCreateDTO
+import dto.applicant.ApplicantReadDTO
+import dto.applicant.ApplicantUpdateDTO
+import dto.assessment.AssessmentReadDTO
+import dto.invite.InviteReadDTO
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.Schema
