+++ management-server/src/main/kotlin/ut/isep/management/controller/SectionController.kt
@@ -1,6 +1,6 @@
 package ut.isep.management.controller
 
-import dto.SectionReadDTO
+import dto.section.SectionReadDTO
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.Schema
