+++ management-server/src/main/kotlin/ut/isep/management/controller/SolutionController.kt
@@ -1,6 +1,6 @@
 package ut.isep.management.controller
 
-import dto.SolutionUpdateDTO
+import dto.assessment.SolutionUpdateDTO
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.Schema
