+++ management-server/src/main/kotlin/ut/isep/management/service/AssignmentService.kt
@@ -1,6 +1,6 @@
 package ut.isep.management.service
 
-import dto.*
+import dto.assignment.AssignmentDTO
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 import ut.isep.management.model.entity.*
