+++ management-server/src/main/kotlin/ut/isep/management/service/InviteService.kt
@@ -1,6 +1,8 @@
 package ut.isep.management.service
 
-import dto.*
+import dto.assessment.AssessmentReadDTO
+import dto.invite.InviteCreateDTO
+import dto.invite.InviteReadDTO
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 import ut.isep.management.model.entity.*
