+++ shared-models/src/main/kotlin/dto/applicant/ApplicantCreateDTO.kt
@@ -1,5 +1,6 @@
-package dto
+package dto.applicant
 
+import dto.CreateDTO
 import enumerable.ApplicantStatus
 import io.swagger.v3.oas.annotations.media.Schema
