+++ shared-models/src/main/kotlin/dto/assessment/InterviewStatusDTO.kt
@@ -1,4 +1,4 @@
-package dto
+package dto.assessment
 
 import io.swagger.v3.oas.annotations.media.Schema
