+++ shared-models/src/main/kotlin/dto/assessment/SolutionUpdateDTO.kt
@@ -1,5 +1,6 @@
-package dto
+package dto.assessment
 
+import dto.assignment.AnswerUpdateDTO
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "A mapping of answer objects to question IDs")
