+++ shared-models/src/main/kotlin/dto/assignment/AnswerUpdateDTO.kt
@@ -1,4 +1,4 @@
-package dto
+package dto.assignment
 
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
