+++ shared-models/src/main/kotlin/dto/invite/InviteCreateDTO.kt
@@ -1,5 +1,6 @@
-package dto
+package dto.invite
 
+import dto.CreateDTO
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "An invite to an applicant")
