+++ shared-models/src/main/kotlin/dto/section/SectionReadDTO.kt
@@ -1,5 +1,7 @@
-package dto
+package dto.section
 
+import dto.ReadDTO
+import dto.assignment.AssignmentDTO
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "A Section, consisting of multiple Assignments")
