+++ shared-models/src/main/kotlin/dto/user/UserCreateDTO.kt
@@ -1,12 +1,12 @@
-package dto
+package dto.user
 
+import dto.CreateDTO
 import enumerable.UserRole
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "A create/read DTO for a user")
-data class UserReadDTO(
-    val id: Long = 0,
+data class UserCreateDTO(
     val name: String = "",
     val email: String = "",
     val role: UserRole? = null
-) : ReadDTO
+) : CreateDTO
